moved unit tests from data to core
